feat(bigtable): wire Diverter on Client and route Open* via TableShim
Three coordinated changes so a follow-up session-data-path patch can
flip traffic through the shim without further Client/OpenTable churn:

1. Client owns a *transport.Diverter, constructed with sessionLoad=0.0
   in NewClientWithConfig. Every call stays on the classic path until a
   future change wires in the session backend and bumps the ratio.

2. Open / OpenTable / OpenAuthorizedView / OpenMaterializedView move
   out of client.go into a new open.go. OpenTable & friends now return
   NewTableShim(classic, nil, c.diverter) — the shim wraps the same
   classic tableImpl as before with routing added on top. Open()
   (returning *Table) stays classic-only for callers that need the
   concrete pointer type.

3. TableShim gains a pickSession() gate: session TableAPI is optional;
   with session==nil pickSession short-circuits before consulting the
   diverter so the pick-count histogram doesn't record session picks
   that got silently downgraded here. ReadRow + Apply route through
   pickSession() instead of calling diverter.UseSession() directly.

Also: DefaultDynamicChannelPoolConfig.MinConns 10 -> 4. Session client
uses a pool of 4 as its footprint default, and today ValidateDynamicConfig
rejects that with "initial connPoolSize (4) must be between MinConns (10)
and MaxConns (200)". Lowering the floor to 4 lets the session pool
validate under the same default config the classic pool uses.

No behavior change on the classic path: sessionLoad=0.0 makes
UseSession() return false in every case, and the pickSession() nil-guard
would defend even if it didn't. All existing bigtable tests pass.
